% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_invalid_chars.R
\name{check_invalid_chars}
\alias{check_invalid_chars}
\title{Check for Invalid Characters in a Data Frame(SDTM)}
\usage{
check_invalid_chars(df, target = text, question = FALSE)
}
\arguments{
\item{df}{A data frame in which to check for invalid characters.}

\item{target}{A symbol specifying the column to check for invalid characters. Default is \code{text}.}

\item{question}{A logical value indicating whether to detect and remove the question mark character. Default is FALSE.}
}
\value{
A data frame with additional columns showing the invalid characters and their transformations.
}
\description{
This function checks for invalid characters in a specified column of a data frame and provides various transformations of the invalid characters.
}
\examples{
library(dplyr)
df <- data.frame(text = c("Hello\x01World", "Example?String"))
check_invalid_chars(df, target = text)
check_invalid_chars(df, target = text, question = TRUE)
}
