% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sdtm_keys.R
\name{check_sdtm_keys}
\alias{check_sdtm_keys}
\title{Check SDTM Keys}
\usage{
check_sdtm_keys(sdtms, dataset_meta)
}
\arguments{
\item{sdtms}{A list of SDTM data.frames.}

\item{dataset_meta}{A data.frame containing dataset and key variables metadata.}
}
\value{
A data.frame with the results of the key checks.
}
\description{
This function checks the key variables in SDTM datasets against the key variables metadata.
}
\examples{
\dontrun{
sdtms <- list(ds1 = data.frame(ID = 1:3), ds2 = data.frame(ID = 1:2))
dataset_meta <- data.frame(dataset = c("ds1", "ds2"), key = c("ID", "ID"))
check_sdtm_keys(sdtms, dataset_meta)
}
}
