% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_dy.R
\name{compute_dy}
\alias{compute_dy}
\title{Compute Day Difference Between Two Dates}
\usage{
compute_dy(refdate, eventdate)
}
\arguments{
\item{refdate}{A reference date, typically the starting point. Can be a character string or Date object.}

\item{eventdate}{An event date, typically the comparison date. Can be a character string or Date object.}
}
\value{
The difference in days between the reference date and the event date. If the difference is non-negative, the result is incremented by 1.
}
\description{
This function calculates the day difference between a reference date and an event date. If the difference is non-negative, the result is incremented by 1 day.
}
\examples{
compute_dy("2023-09-15", "2023-09-20")
}
