% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_iso8601.R
\name{conv_iso8601}
\alias{conv_iso8601}
\title{Convert character date and time to ISO 8601 format}
\usage{
conv_iso8601(datetime, time = NA, type = "DMY")
}
\arguments{
\item{datetime}{A character vector representing the date or datetime.}

\item{time}{A character vector representing the time (optional).}

\item{type}{A character to specify the format of the date ("DMY", "YMD", "MDY", "YDM").}
}
\value{
A character vector in ISO 8601 format.
}
\description{
This function takes a datetime string and optionally a time string, and converts them to ISO 8601 format.
The type parameter controls the order of year, month, and day.
}
