% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_iso8601.R
\name{conv_iso8601_num_d}
\alias{conv_iso8601_num_d}
\title{Convert Date to ISO 8601 Format (Date)}
\usage{
conv_iso8601_num_d(date)
}
\arguments{
\item{date}{A Date object or character string.}
}
\value{
A character string in ISO 8601 format ("YYYY-MM-DD") if the input is a Date, otherwise returns the input as a character string.
}
\description{
This function converts a Date object to an ISO 8601 formatted date string ("YYYY-MM-DD").
}
\examples{
conv_iso8601_num_d(as.Date("2023-09-15"))
conv_iso8601_num_d("2023-09-15")
}
