% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_iso8601.R
\name{conv_iso8601_num_hms}
\alias{conv_iso8601_num_hms}
\title{Convert POSIXct to ISO 8601 Format (Date and Time with Seconds)}
\usage{
conv_iso8601_num_hms(date)
}
\arguments{
\item{date}{A POSIXct object or character string.}
}
\value{
A character string in ISO 8601 format ("YYYY-MM-DDTHH:MM:SS") if the input is a POSIXct object, otherwise returns the input as a character string.
}
\description{
This function converts a POSIXct object to an ISO 8601 formatted datetime string ("YYYY-MM-DDTHH:MM:SS"), including seconds.
}
\examples{
conv_iso8601_num_hms(as.POSIXct("2023-09-15 14:30:45"))
conv_iso8601_num_hms("2023-09-15T14:30:45")
}
