% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_iso8601.R
\name{derive_dtf_from_dtc}
\alias{derive_dtf_from_dtc}
\title{Derive DTF (Date Imputation Flag) Variable from DTC}
\usage{
derive_dtf_from_dtc(dtc, null_all_na = TRUE)
}
\arguments{
\item{dtc}{A character vector containing datetime strings. Each string should follow the ISO 8601 format (e.g., \code{YYYY}, \code{YYYY-MM}, \code{YYYY-MM-DD}).}

\item{null_all_na}{A logical value indicating whether to return \code{NULL} if all results are \code{NA}. Defaults to \code{TRUE}.}
}
\value{
A character vector indicating the level of date imputation:
\itemize{
\item \code{"M"} for month-level imputation (e.g., \code{YYYY} -> \code{YYYY-MM-DD}).
\item \code{"D"} for day-level imputation (e.g., \code{YYYY-MM} -> \code{YYYY-MM-DD}).
If \code{null_all_na} is \code{TRUE} and all results are \code{NA}, the function returns \code{NULL}.
}
}
\description{
This function creates the DTF (Date Imputation Flag) variable for ADaM datasets.
It determines the level of date imputation (e.g., year, month, or day) based on the length of the input datetime string (\code{dtc}).
}
\examples{
# Example with valid datetime strings
derive_dtf_from_dtc(c("2023", "2023-05", "2023-05-15"))
# Returns: c("D", "M", NA)

# Example with non-inputed dates
derive_dtf_from_dtc(c("2023-01-01", "2023-05-01", "2023-05-15"))
# Returns: NULL (if all results are NA after processing)

# Example with non-inputed dates with null_all_na = FALSE
derive_dtf_from_dtc(c("2023-01-01", "2023-05-01", "2023-05-15"))
# Returns: c(NA, NA, NA)

}
