% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_iso8601.R
\name{derive_tmf_from_dtc}
\alias{derive_tmf_from_dtc}
\title{Derive TMF (Time Imputation Flag) Variable from DTC}
\usage{
derive_tmf_from_dtc(dtc, null_all_na = TRUE)
}
\arguments{
\item{dtc}{A character vector containing datetime strings. Each string should follow the ISO 8601 format (e.g., \code{YYYY-MM-DDTHH}, \code{YYYY-MM-DDTHH:MM}, \code{YYYY-MM-DDTHH:MM:SS}).}

\item{null_all_na}{A logical value indicating whether to return \code{NULL} if all results are \code{NA}. Defaults to \code{TRUE}.}
}
\value{
A character vector indicating the level of time imputation:
\itemize{
\item \code{"H"} for hour-level imputation (e.g., \code{YYYY-MM-DD} -> \code{YYYY-MM-DDTHH:MM:SS}).
\item \code{"M"} for minute-level imputation (e.g., \code{YYYY-MM-DDTHH} -> \code{YYYY-MM-DDTHH:MM:SS}).
\item \code{"S"} for second-level imputation (e.g., \code{YYYY-MM-DDTHH:MM} -> \code{YYYY-MM-DDTHH:MM:SS}).
\item \code{NA} for non-imputation (e.g., \code{YYYY-MM-DDTHH:MM:SS} -> \code{YYYY-MM-DDTHH:MM:SS}).
If \code{null_all_na} is \code{TRUE} and all results are \code{NA}, the function returns \code{NULL}.
}
}
\description{
This function creates the TMF (Time Imputation Flag) variable for ADaM datasets.
It determines the level of time imputation (e.g., hour, minute, or second) based on the length of the input datetime string (\code{dtc}).
}
\examples{
# Example with valid datetime strings
derive_tmf_from_dtc(c("2023", "2023-05-15", "2023-05-15T12:30", "2023-05-15T12:30:45"))
# Returns: c("H", "H", "S", NA)

# Example with non-imputed times
derive_tmf_from_dtc(c("2023-01-01T12:00:00", "2023-05-01T12:30:00", "2023-05-15T12:30:45"))
# Returns: NULL (if all results are NA after processing)

# Example with non-imputed times and null_all_na = FALSE
derive_tmf_from_dtc(c("2023-01-01T12:00:00", "2023-05-01T12:30:00", "2023-05-15T12:30:45"), null_all_na = FALSE)
# Returns: c(NA, NA, NA)

}
