% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_invalid_chars.R
\name{detect_invalid_chars}
\alias{detect_invalid_chars}
\title{Detect Invalid Characters for SDTM (non-ascii + alpha)}
\usage{
detect_invalid_chars(text, question = FALSE)
}
\arguments{
\item{text}{A character string from which invalid characters are to be removed.}

\item{question}{A logical value indicating whether to detect and remove the question mark character. Default is FALSE.}
}
\value{
A character string with invalid characters removed.
}
\description{
This function detects invalid characters for SDTM from a given text. It can optionally detect the question mark character, it can be used for broken characters on SAS.
}
\examples{
detect_invalid_chars("Hello\x01World")
detect_invalid_chars("Hello?World", question = TRUE)
}
