% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ct_decode.R
\name{get_ct_decode}
\alias{get_ct_decode}
\title{Get Decode Value from CT Metadata}
\usage{
get_ct_decode(ct_meta, id, term, reverse = FALSE)
}
\arguments{
\item{ct_meta}{A data.frame containing the metadata with columns \code{id}, \code{term}, and \code{decode}. read_sdtm_metadata_p21() can be used to read the codelists metadata from a P21 specification file.}

\item{id}{A single character string representing the ID.}

\item{term}{A character vector representing the term(s).}

\item{reverse}{A logical value indicating whether to reverse the lookup (default is FALSE).}
}
\value{
A character vector of decode values corresponding to the specified ID and term.
}
\description{
This function retrieves the decode value from a metadata data.frame based on the specified ID and term.
}
\examples{
library(dplyr)

# Create a sample metadata data.frame
ct_meta <- data.frame(
  id = c("LATESTCD", "LBTESTCD", "LBTESTCD"),
  term = c("RBC", "MONO", "LYM"),
  decode = c("Erythrocyte Count", "Monocyte Count", "Lymphocyte Count"),
  stringsAsFactors = FALSE
)

df <- data.frame(USUBJID = rep(1:3), LBTESTCD = c("RBC", "MONO", "LYM"))

# Get decode value
df |> mutate(LBTEST = get_ct_decode(ct_meta, "LBTESTCD", LBTESTCD))

df2 <- data.frame(USUBJID = rep(1:3), LBTEST = c("Erythrocyte Count", "Monocyte Count", "Lymphocyte Count"))

# Get decode value with reverse lookup
df2 |> mutate(LBTEST = get_ct_decode(ct_meta, "LBTESTCD", LBTESTCD, reverse = TRUE))

}
