% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ct_decode.R
\name{get_ct_term}
\alias{get_ct_term}
\title{Get Term Value from CT Metadata}
\usage{
get_ct_term(ct_meta, id, decode)
}
\arguments{
\item{ct_meta}{A \code{data.frame} containing the metadata with columns \code{id}, \code{term}, and \code{decode}. The metadata can be read using \code{read_sdtm_metadata_p21()} from a P21 specification file.}

\item{id}{A single character string representing the ID.}

\item{term}{A character vector representing the decode value(s).}
}
\value{
A character vector of term values corresponding to the specified ID and decode value(s).
}
\description{
This function retrieves the term value from a metadata \code{data.frame} based on the specified ID and decode value.
}
\examples{
library(dplyr)

# Create a sample metadata data.frame
ct_meta <- data.frame(
  id = c("LATESTCD", "LBTESTCD", "LBTESTCD"),
  term = c("RBC", "MONO", "LYM"),
  decode = c("Erythrocyte Count", "Monocyte Count", "Lymphocyte Count"),
  stringsAsFactors = FALSE
)

df <- data.frame(USUBJID = rep(1:3), LBTEST = c("Erythrocyte Count", "Monocyte Count", "Lymphocyte Count"))

# Get term value
df |> mutate(LBTESTCD = get_ct_term(ct_meta, "LBTESTCD", LBTEST))

}
