% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_column_by_max_bytes.R
\name{get_split_var_names}
\alias{get_split_var_names}
\title{Generate split variable names based on a base variable name}
\usage{
get_split_var_names(var_name, len, max_len = len)
}
\arguments{
\item{var_name}{The base name of the variable to be split.}

\item{len}{The number of split variables to generate names for.}

\item{max_len}{The maximum length of the variable names. Default is the length of the base variable name.}
}
\value{
A character vector containing the generated variable names.
}
\description{
This function generates a list of variable names for split columns based on a base variable name.
The names are generated to ensure that they fit within a specified maximum length.
}
\examples{
var_names <- get_split_var_names("variable", 5)
print(var_names)
}
