% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sdtm_keys.R
\name{is_unique_key}
\alias{is_unique_key}
\title{Check if Key is Unique}
\usage{
is_unique_key(df, key)
}
\arguments{
\item{df}{A data.frame to check.}

\item{key}{A character vector of key variables.}
}
\value{
TRUE if the key is unique, FALSE otherwise.
}
\description{
This function checks if the key variables form a unique key in the data.frame.
}
\examples{
df <- data.frame(ID = c(1, 2, 2), NAME = c("A", "B", "B"))
is_unique_key(df, c("ID", "NAME"))
}
