% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_supp_df.R
\name{make_supp_df}
\alias{make_supp_df}
\title{Create a suppqual DataFrame}
\usage{
make_supp_df(df, suppmeta, idver = NULL)
}
\arguments{
\item{df}{DataFrame. The original sdtm dataframe with supp variables.}

\item{suppmeta}{DataFrame. The dataframe containing supp variables metadata. The suppmeta have to have the following columns: RDOMAIN, IDVAR, QNAM, QLABEL, QORIG, QEVAL.}

\item{idver}{String. To replace IDVAR provided in suppmeta. (optional).}
}
\value{
DataFrame. The supplementary dataframe.
}
\description{
This function creates a suppqual dataframe using the specified dataframe and suppqual metadata dataframe. AP domains are not supported.
}
\examples{
dm <- data.frame(
  STUDYID = c("HOGE-01", "HOGE-01", "HOGE-01"),
  DOMAIN = c("DM", "DM", "DM"),
  USUBJID = c("HOGE-01-001", "HOGE-01-002", "HOGE-01-003"),
  SUBJID = c("001", "002", "003"),
  SUPP1 = c("ASIAN", "ASIAN", ""),
  SUPP2 = c("WHITE", "", ""),
  SUPP3 = c("", "", "BLACK OR AFRICAN AMERICAN")
)

ae <- data.frame(
  STUDYID = c("HOGE-01", "HOGE-01", "HOGE-01"),
  DOMAIN = c("AE", "AE", "AE"),
  USUBJID = c("HOGE-01-001", "HOGE-01-002", "HOGE-01-003"),
  AESEQ = c(1, 20, 30),
  AETERM = c("term1", "term2", "term3"),
  SUPP1 = c("supp1a", "supp1b", ""),
  SUPP2 = c("supp2a", "", ""),
  SUPP3 = c("supp3a", "supp3b", NA)
)

apcm <- data.frame(
  STUDYID = c("HOGE-01", "HOGE-01", "HOGE-01"),
  DOMAIN = c("APCM", "APCM", "APCM"),
  APID = c("HOGE-01-001-P", "HOGE-01-002-P", "HOGE-01-003-P"),
  CMSPID = c("CM-001", "CM-002", "CM-003"),
  CMTRT = c("trt1", "trt2", "trt3"),
  SUPP1 = c("supp1a", "supp1b", "supp1c"),
  SUPP2 = c("supp2a", "supp2b", ""),
  SUPP3 = c("", "", "")
)

suppmeta <- data.frame(
  RDOMAIN = c("DM", "DM", "DM", "AE", "AE", "AE", "APCM", "APCM", "APCM"),
  IDVAR = c("", "", "", "AESEQ", "AESEQ", "AESEQ", "CMSPID", "CMSPID", "CMSPID"),
  QNAM = c("SUPP1", "SUPP2", "SUPP3", "SUPP1", "SUPP2", "SUPP3", "SUPP1", "SUPP2", "SUPP3"),
  QLABEL = c("Supp Label 1", "Supp Label 2", "Supp Label 3", "Supp Label 1", "Supp Label 2", "Supp Label 3", "Supp Label 1", "Supp Label 2", "Supp Label 3"),
  QORIG = c("CRF", "CRF", "CRF", "CRF", "CRF", "CRF", "CRF", "CRF", "CRF"),
  QEVAL = c("", "", "", "", "", "", "", "", "")
)

library(tidyverse)
suppdm <- make_supp_df(dm, suppmeta)
suppae <- make_supp_df(ae, suppmeta)
suppcm <- make_supp_df(apcm, suppmeta)
}
