% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sas_datasets.R
\name{pickup_subjects}
\alias{pickup_subjects}
\title{Filter Subjects from a List of Data Frames}
\usage{
pickup_subjects(df_list, varname, varval = NA)
}
\arguments{
\item{df_list}{A list of data frames to be filtered.}

\item{varname}{A character string specifying the name of the variable (column) to filter by.}

\item{varval}{A vector of values used to filter the rows of the data frames. If \code{NA}, no filtering is applied. Default is \code{NA}.}
}
\value{
A list of filtered data frames. If a data frame does not contain the specified variable, it is returned unmodified.
}
\description{
This function filters subjects (rows) from a list of data frames based on a specified variable and its value(s).
If the variable is present in the data frame, the rows where the variable matches the specified values are retained.
If no values are provided for filtering, the function returns the original list of data frames.
}
\examples{
\dontrun{
# Example with filtering based on a variable
filtered_dfs <- pickup_subjects(df_list, "subject_id", c("101", "102"))

# Example without filtering (returns the original list)
all_dfs <- pickup_subjects(df_list, "subject_id")
}
}
