% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_metadata.R
\name{read_sdtm_meta_supp}
\alias{read_sdtm_meta_supp}
\title{Read SDTM Supplemental Qualifiers Metadata from Excel File}
\usage{
read_sdtm_meta_supp(
  spec,
  supp_sheet = "Suppqual",
  dataset_col = "Dataset",
  qnam_col = "QNAM",
  qlabel_col = "QLABEL",
  qevel_col = "QEVAL",
  qorig_col = "QORIG",
  idvar_col = "IDVAR"
)
}
\arguments{
\item{spec}{A character string specifying the path to the Excel file.}

\item{supp_sheet}{A character string specifying the name of the sheet containing the supplemental qualifiers metadata. Default is "Suppqual".}

\item{dataset_col}{A character string specifying the name of the column containing dataset names. Default is "Dataset".}

\item{qnam_col}{A character string specifying the name of the column containing QNAM values. Default is "QNAM".}

\item{qlabel_col}{A character string specifying the name of the column containing QLABEL values. Default is "QLABEL".}

\item{qevel_col}{A character string specifying the name of the column containing QEVAL values. Default is "QEVAL".}

\item{qorig_col}{A character string specifying the name of the column containing QORIG values. Default is "QORIG".}

\item{idvar_col}{A character string specifying the name of the column containing IDVAR values. Default is "IDVAR".}
}
\value{
A data frame containing the supplemental qualifiers metadata.
}
\description{
This function reads SDTM supplemental qualifiers metadata from a specified sheet in an Excel file.
}
