% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sas_datasets.R
\name{read_xpts}
\alias{read_xpts}
\title{Read XPT Files from a Folder}
\usage{
read_xpts(folder, xpts = NA)
}
\arguments{
\item{folder}{A character string specifying the path to the folder containing the XPT files.}

\item{xpts}{A character vector specifying the names of the XPT files to read (without the .xpt extension). If NA, all XPT files in the folder are read. Default is NA.}
}
\value{
A named list of data frames, where each data frame corresponds to an XPT file.
}
\description{
This function reads XPT files from a specified folder and returns them as a list of data frames. Optionally, a subset of XPT files can be specified.
}
\examples{
\dontrun{
# Read all XPT files from a folder
data_list <- read_xpts("path/to/folder")

# Read specific XPT files from a folder
data_list <- read_xpts("path/to/folder", xpts = c("file1", "file2"))
}
}
