% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_invalid_chars.R
\name{replace_control_chars}
\alias{replace_control_chars}
\title{Replace Control Characters with Descriptive Names}
\usage{
replace_control_chars(text)
}
\arguments{
\item{text}{A character string in which control characters are to be replaced.}
}
\value{
A character string with control characters replaced by their descriptive names.
}
\description{
This function replaces control characters in a given text with their descriptive names.
}
\examples{
replace_control_chars("Hello\x01World")
replace_control_chars("Example\x0AString")
}
