% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_column_by_max_bytes.R
\name{set_label_on_sep_columns}
\alias{set_label_on_sep_columns}
\title{Set a label on split columns based on specified column in a data frame}
\usage{
set_label_on_sep_columns(df, var_name, label = attr(df[[var_name]], "label"))
}
\arguments{
\item{df}{A data frame containing the columns to be labeled.}

\item{var_name}{The base name of the variable to be labeled.}

\item{label}{A character string specifying the label to be set. Must be less than 40 characters. Default is the label attribute of the first specified column.}
}
\value{
The data frame with the label attribute set on the specified columns.
}
\description{
This function sets a label on split columns based on specified column in a data frame. The columns are identified based on a base variable name followed by one or two digits.
}
\examples{
df <- data.frame(
  id = 1:3,
  text1 = c("This is a test.", "Another test.", "Short text"),
  text2 = c("More text.", "Even more text.", "Tiny text")
)
df <- set_label_on_sep_columns(df, "text", label = "Example Label")
print(attr(df$text1, "label"))
print(attr(df$text2, "label"))
}
