% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sdtm_keys.R
\name{split_key_text}
\alias{split_key_text}
\title{Split Key Text}
\usage{
split_key_text(key_text)
}
\arguments{
\item{key_text}{A character string containing key variables separated by delimiters.}
}
\value{
A character vector of key variables.
}
\description{
This function splits a key text into individual key components.
}
\examples{
split_key_text("ID, NAME")
}
