% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_column_by_max_bytes.R
\name{split_text_by_max_bytes}
\alias{split_text_by_max_bytes}
\title{Split a text into multiple parts based on maximum byte length}
\usage{
split_text_by_max_bytes(text, max_bytes = 200, splitter = " |;|,|\\\\.|-")
}
\arguments{
\item{text}{A character string to be split.}

\item{max_bytes}{An integer specifying the maximum byte length for each split. Default is 200.}

\item{splitter}{A regular expression pattern used to identify split points. Default is " |;|,|\\.".}
}
\value{
A character vector with the text split into parts.
}
\description{
This function splits a given text into multiple parts, ensuring that each part does not exceed the specified maximum byte length. It also ensures that the text is not split in the middle of a multibyte character.
}
\examples{
text <- "This is a very long text that needs to be split into smaller parts."
result <- split_text_by_max_bytes(text, max_bytes = 10)
print(result)
}
