% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_width.R
\name{text_width_courier_new}
\alias{text_width_courier_new}
\title{Calculate the Width of a String in Courier New Font}
\usage{
text_width_courier_new(text, points = 12, bold = FALSE)
}
\arguments{
\item{text}{A character string whose width is to be estimated.}

\item{points}{A numeric value representing the font size in points. Default is 12.}

\item{bold}{A logical value indicating whether the text is bold. Default is FALSE.}
}
\value{
A numeric value representing the estimated width of the string in points.
}
\description{
This function estimates the width of a given string when rendered in Courier New font at a specified point size.
}
\examples{
text_width_courier_new("AETERM")
text_width_courier_new("AE (Adverse Events)", points = 16, bold = TRUE)
}
