% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_invalid_chars.R
\name{to_hex}
\alias{to_hex}
\title{Convert a String to Hexadecimal Representation}
\usage{
to_hex(text)
}
\arguments{
\item{text}{A character string to be converted to hexadecimal representation.}
}
\value{
A character vector where each element is the hexadecimal representation of the corresponding character in the input string.
}
\description{
This function converts each character in a string to its hexadecimal representation.
}
\examples{
to_hex("Hello")
to_hex("World")
}
