% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_char_date.R
\name{comp_char_dates}
\alias{comp_char_dates}
\title{Compare Character Date Vectors}
\usage{
comp_char_dates(x, y, mark = "<=")
}
\arguments{
\item{x}{A character vector representing the first set of dates.}

\item{y}{A character vector representing the second set of dates.}

\item{mark}{A character string representing the comparison operator.
Must be one of "<", "<=", ">", ">=", "==", "!=". Default is "<=".}
}
\value{
A logical vector indicating the result of the comparison for each element, or an error if the input vectors have different lengths.
}
\description{
This function compares two character date vectors element-wise based on a specified comparison operator.
}
\examples{
x <- c("2023-10-01", "2023-10-02")
y <- c("2023-10-01", "2023-10-03")
comp_char_dates(x, y, "<=") # TRUE, TRUE
comp_char_dates(x, y, ">")  # FALSE, FALSE

}
