% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2rtf_tools.R
\name{convert_units}
\alias{convert_units}
\title{Convert Units}
\usage{
convert_units(values, from = "mm", to = "mm")
}
\arguments{
\item{values}{Numeric vector of values}

\item{from}{Source unit}

\item{to}{Target unit}
}
\value{
Converted values
}
\description{
Convert Units
}
